+++ app/Modules/Invoices/Controllers/InvoiceEditController.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of BillingTrack.
  *
- * 
+ *
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -19,6 +19,7 @@ use BT\Modules\Invoices\Models\InvoiceItem;
 use BT\Modules\Invoices\Support\InvoiceTemplates;
 use BT\Modules\Invoices\Requests\InvoiceUpdateRequest;
 use BT\Modules\ItemLookups\Models\ItemLookup;
+use BT\Modules\Products\Models\Product;
 use BT\Modules\TaxRates\Models\TaxRate;
 use BT\Support\DateFormatter;
 use BT\Support\Statuses\InvoiceStatuses;
@@ -46,12 +47,14 @@ class InvoiceEditController extends Controller
     public function update(InvoiceUpdateRequest $request, $id)
     {
         // Unformat the invoice dates.
-        $invoiceInput                 = $request->except(['items', 'custom', 'apply_exchange_rate']);
+        $invoiceInput = $request->except(['items', 'custom', 'apply_exchange_rate']);
         $invoiceInput['invoice_date'] = DateFormatter::unformat($invoiceInput['invoice_date']);
-        $invoiceInput['due_at']       = DateFormatter::unformat($invoiceInput['due_at']);
+        $invoiceInput['due_at'] = DateFormatter::unformat($invoiceInput['due_at']);
 
         // Save the invoice.
         $invoice = Invoice::find($id);
+        $oldstatus = $invoice->invoice_status_id;
+        $newstatus = $invoiceInput['invoice_status_id'];
         $invoice->fill($invoiceInput);
         $invoice->save();
 
@@ -59,19 +62,28 @@ class InvoiceEditController extends Controller
         $invoice->custom->update(request('custom', []));
 
         // Save the items.
-        foreach ($request->input('items') as $item)
-        {
+        foreach ($request->input('items') as $item) {
             $item['apply_exchange_rate'] = request('apply_exchange_rate');
 
-            if (!isset($item['id']) or (!$item['id']))
-            {
+            if (!isset($item['id']) or (!$item['id'])) {
                 $saveItemAsLookup = $item['save_item_as_lookup'];
                 unset($item['save_item_as_lookup']);
 
-                InvoiceItem::create($item);
+                $newitem = InvoiceItem::create($item);
+                // product numstock update
+                // if inv tracking is on and invoice is sent
+                // and item is tracked inventory, decrement onhand
+                if (config('bt.updateInvProductsDefault') && $newitem->invoice->status_text == 'sent') {
+                    if ($newitem->resource_id && $newitem->resource_table == 'products'
+                        && $newitem->product()->tracked()->get()->isNotEmpty()) {
+                        $newitem->product->numstock -= $newitem->quantity;
+                        $newitem->product->save();
+                        $newitem->is_tracked = 1;
+                        $newitem->save();
+                    }
+                }
 
-                if ($saveItemAsLookup)
-                {
+                if ($saveItemAsLookup) {
                     ItemLookup::create([
                         'name'          => $item['name'],
                         'description'   => $item['description'],
@@ -80,10 +92,54 @@ class InvoiceEditController extends Controller
                         'tax_rate_2_id' => $item['tax_rate_2_id'],
                     ]);
                 }
-            }
-            else
-            {
+            } else {
                 $invoiceItem = InvoiceItem::find($item['id']);
+                $qtydiff = abs($invoiceItem->quantity - $item['quantity']);
+                // product numstock updating
+                if (config('bt.updateInvProductsDefault')) {
+                    // if quantity changed
+                    if ($qtydiff && $invoiceItem->is_tracked) {
+                        switch ($invoiceItem->quantity <=> $item['quantity']) {
+                            case 0:
+                                break;
+                            case -1:
+                                // decrement numstock
+                                $product = Product::find($invoiceItem->resource_id);
+                                $product->numstock -= $qtydiff;
+                                $product->save();
+                                break;
+                            case 1:
+                                // increment numstock
+                                $product = Product::find($invoiceItem->resource_id);
+                                $product->numstock += $qtydiff;
+                                $product->save();
+                                break;
+                        }
+                    }
+
+                    // if status changed to sent from draft or canceled
+                    if (($oldstatus == 1 || $oldstatus == 4) && $newstatus == 2) {
+                        // if item has NOT already been tracked
+                        // and item is tracked inventory, decrement onhand
+                        if ($invoiceItem->resource_id && !$invoiceItem->is_tracked && $invoiceItem->resource_table == 'products'
+                            && $invoiceItem->product()->tracked()->get()->isNotEmpty()) {
+                            $invoiceItem->product->numstock -= $item['quantity'];
+                            $invoiceItem->product->save();
+                            $invoiceItem->is_tracked = 1;
+                        }
+                    }
+                    //if status changed from sent to draft or canceled
+                    if ($oldstatus == 2 && ($newstatus == 1 || $newstatus == 4)) {
+                        // if item has already been tracked
+                        // and item is tracked inventory, increment onhand
+                        if ($invoiceItem->resource_id && $invoiceItem->is_tracked && $invoiceItem->resource_table == 'products'
+                            && $invoiceItem->product()->tracked()->get()->isNotEmpty()) {
+                            $invoiceItem->product->numstock += $invoiceItem->quantity;
+                            $invoiceItem->product->save();
+                            $invoiceItem->is_tracked = 0;
+                        }
+                    }
+                }
                 $invoiceItem->fill($item);
                 $invoiceItem->save();
             }