+++ app/Modules/Invoices/Controllers/InvoiceItemController.php
@@ -13,11 +13,19 @@ namespace BT\Modules\Invoices\Controllers;
 
 use BT\Http\Controllers\Controller;
 use BT\Modules\Invoices\Models\InvoiceItem;
+use BT\Modules\Products\Models\Product;
 
 class InvoiceItemController extends Controller
 {
     public function delete()
     {
+        $invoiceitem = InvoiceItem::find(request('id'));
+
+        if (config('bt.updateInvProductsDefault') && $invoiceitem->is_tracked){
+            $product = Product::find($invoiceitem->resource_id);
+            $product->numstock += $invoiceitem->quantity;
+            $product->save();
+        }
         InvoiceItem::destroy(request('id'));
     }
 }