+++ app/Modules/Products/Views/products/_table.blade.php
@@ -28,9 +28,17 @@
             <td>{{ $product->inventorytype->name }}</td>
             <td>{{ $product->numstock }}</td>
             <td>{{ $product->active }}</td>
-            <td> <a href="{{ route('products.edit', [$product->id]) }}" class="btn btn-primary btn-sm "><i
-                            class="fa fa-edit"></i>
-                    @lang('bt.edit') </a></td>
+            <td>
+            <button type="button" class="btn btn-secondary btn-sm dropdown-toggle" data-toggle="dropdown">
+                @lang('bt.options')
+            </button>
+            <div class="dropdown-menu dropdown-menu-right" role="menu">
+                <a class="dropdown-item" href="{{ route('products.edit', [$product->id]) }}"><i
+                            class="fa fa-edit"></i> @lang('bt.edit')</a>
+                <a href ="javascript:void(0)" class="create-purchaseorder dropdown-item" data-name="{{ $product->vendor->name }}"
+                   data-productid="{{ $product->id }}"><i class="fa fa-shopping-cart"></i> @lang('bt.order')</a>
+            </div>
+            </td>
         </tr>
     @endforeach
     </tbody>