+++ CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added config for collapsed side menu
 - corrected invoice update when payment amount is changed.
 - corrected payment emailing
-- added product inventory tracking ability.
+- added product inventory tracking ability to sent invoices.
   note:previously unused "type" column of products table has been changed to "inventorytype_id"
   Any existing entries in the type column will be migrated to the inventory_types table and updated
   in the product. This table is currently not editable within the application.